// app/backend/src/interfaces/IServiceBoards.ts
export type TBoards = {
  name: string;
  totalPoints: number;
  totalGames: number;
  totalVictories: number;
  totalDraws: number;
  totalLosses: number;
  goalsFavor: number;
  goalsOwn: number;
  goalsBalance: number;
  efficiency: string;

};

export default interface IServiceBoards {
  find(): Promise<TBoards[]>
}